refactor: Mock API BASE_URL 추가
