+++ client/src/features/FollowingTab/apis/getFollowings.ts
@@ -1,6 +1,11 @@
+import { API_URL } from '@constants/url';
 import axios from 'axios';
 
+const BASE_URL = `${API_URL}/api`;
+
 const getFollowings = (query: string) =>
-  axios.get(`/api/user/followings?query=${query}`).then((res) => res.data);
+  axios
+    .get(`${BASE_URL}/user/followings?query=${query}`)
+    .then((res) => res.data);
 
 export default getFollowings;
