+++ client/src/mocks/handlers/Friend.js
@@ -1,28 +1,34 @@
+import { API_URL } from '@constants/url';
 import { rest } from 'msw';
 
-import { users } from '../data';
+import { users } from '../data/users';
 
-const GetFollowings = rest.get('/api/user/followings', (req, res, ctx) => {
-  const query = req.url.searchParams.get('query') ?? '';
+const BASE_URL = `${API_URL}/api`;
 
-  return res(
-    ctx.delay(),
-    ctx.status(200),
-    ctx.json({
-      statusCode: 200,
-      result: {
-        followings: query
-          ? users.filter(({ nickname }) =>
-              nickname.toUpperCase().includes(query.toUpperCase()),
-            )
-          : users,
-      },
-    }),
-  );
-});
+const GetFollowings = rest.get(
+  `${BASE_URL}/user/followings`,
+  (req, res, ctx) => {
+    const query = req.url.searchParams.get('query') ?? '';
+
+    return res(
+      ctx.delay(),
+      ctx.status(200),
+      ctx.json({
+        statusCode: 200,
+        result: {
+          followings: query
+            ? users.filter(({ nickname }) =>
+                nickname.toUpperCase().includes(query.toUpperCase()),
+              )
+            : users,
+        },
+      }),
+    );
+  },
+);
 
 const UpdateFollowing = rest.post(
-  '/api/user/following/:userId',
+  `${BASE_URL}/user/following/:userId`,
   (req, res, ctx) => {
     const { userId } = req.params;
     const idx = users.findIndex((user) => user._id === userId);
